Change to a persistent database connection
Refactor the DBUtil class into a singleton for the database connection 
named DBManager. Create an instance of the connection at the beginning 
of the main method. Close the connection at the end of the main method. 
Remove all the older Connection objects in the GuestController class.

Add access to the hotel read functionality.