Add complete functionality for more tables
Add  functionality for the hotel and manager tables. Refactor the 
AppMenu class to move the general record type selection menu to the 
DBManager class to allow the same menu to be generated with the action 
the user requested to be performed passed to it. Clean up the AppMenu 
class and do other minor cleanup.