Convert project to GUI
Convert project to GUI and run it using the new Hotela class. Add full 
read and update functionality for the five main tables. Add a 
constructor for each bean with everything but the class ID. Add a SQL 
query to get only the ID for each table. Rename the main package from 
com.shastri.hotelaccommodation to me.shastri.hotelaccommodation.